bot-service/requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ── LangChain + Providers ────────────────────────────────────────────────────
langchain==0.3.7
langchain-openai==0.2.8
langchain-google-genai
langchain-anthropic
langchain-core==0.3.18

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0          # SQLite driver for tests only

# ── Auth ─────────────────────────────────────────────────────────────────────
PyJWT==2.9.0

# ── Config ───────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
